refactor(client): separate pages and components
