+++ packages/client/src/App.tsx
@@ -4,7 +4,7 @@ import { Configuration } from '@fullstack/sdk';
 import { CssBaseline } from '@mui/material';
 import { createTheme, ThemeProvider } from '@mui/material/styles';
 
-import { Home } from './Home';
+import { Home } from './pages';
 
 const theme = createTheme({
   palette: {
@@ -30,6 +30,7 @@ const DEV_API_BASE_PATH = 'http://localhost:4000';
 const mode = `import.meta.env.MODE` as string;
 const isProduction = mode === `"production"`;
 
+// TODO: Use the API
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 const configuration = new Configuration({
   basePath: isProduction ? PROD_API_BASE_PATH : DEV_API_BASE_PATH,
