+++ packages/client/src/assets/hero.png
