+++ packages/client/src/components/index.ts
@@ -1,0 +1,2 @@
+export * from './AppNavBar';
+export * from './Hero';
