+++ packages/client/src/pages/Home.tsx
@@ -1,0 +1,10 @@
+import { AppNavBar, Hero } from '../components';
+
+const Home = () => (
+  <>
+    <AppNavBar />
+    <Hero />
+  </>
+);
+
+export { Home };
